Move InputSolver to services
