+++ app/services/InputSolver.coffee
