+++ app/services/InputSolver.js
