+++ tests/InputSolverTest.js
@@ -1,7 +1,7 @@
 // Generated by CoffeeScript 1.9.3
 var InputSolver, expect;
 
-InputSolver = require('../app/InputSolver');
+InputSolver = require('../app/services/InputSolver');
 
 expect = chai.expect;
